+++ examples/mnist/task.py
@@ -14,7 +14,7 @@ class MNIST(Task):
 			input_key="image",
 			label_key="label",
 			input_dtype=tf.float32,
-			label_dtype=tf.int64
+			label_dtype=tf.int32
 		)
 	
 	def dataset_partitioner(self) -> DatasetPartitioner:
